Device model changed, refactor device duplicate service
