+++ IntuneAssistant.Cli/Commands/Devices/DeviceDuplicateCmd.cs
@@ -1,5 +1,6 @@
 using System.CommandLine;
 using IntuneAssistant.Infrastructure.Interfaces;
+using IntuneAssistant.Infrastructure.Interfaces.Devices;
 using IntuneAssistant.Infrastructure.Services;
 using IntuneAssistant.Models;
 using IntuneAssistant.Models.Devices;
@@ -36,18 +37,18 @@ public class FetchDeviceDuplicateCommandOptions : ICommandOptions
 
 public class FetchDeviceDuplicateCommandHandler : ICommandOptionsHandler<FetchDeviceDuplicateCommandOptions>
 {
-    private readonly IDeviceDuplicateService _deviceDuplicateService;
+    private readonly IDeviceService _deviceService;
 
-    public FetchDeviceDuplicateCommandHandler(IDeviceDuplicateService deviceDuplicateService)
+    public FetchDeviceDuplicateCommandHandler(IDeviceService deviceService)
     {
-        _deviceDuplicateService = deviceDuplicateService;
+        _deviceService = deviceService;
     }
 
     public async Task<int> HandleAsync(FetchDeviceDuplicateCommandOptions options)
     {
 
         var removeProvided = options.Remove;
-        var forceProviced = options.Force;
+        var forceProvided = options.Force;
         var accessToken = await new IdentityHelperService().GetAccessTokenSilentOrInteractiveAsync();
         if (string.IsNullOrWhiteSpace(accessToken))
         {
@@ -74,7 +75,7 @@ public class FetchDeviceDuplicateCommandHandler : ICommandOptionsHandler<FetchDe
             .StartAsync("Fetching duplicate devices from Intune", async _ =>
             {
 
-                    var allDeviceResults = await _deviceDuplicateService.GetDuplicateDevicesListAsync(accessToken, deviceFilterOptions, exportOptions);
+                    var allDeviceResults = await _deviceService.GetDuplicateDevicesListAsync(accessToken, deviceFilterOptions, exportOptions);
                     if (allDeviceResults is not null)
                     {
                         devices.AddRange(allDeviceResults.Select(x => x.ToDeviceModel()));
@@ -102,9 +103,9 @@ public class FetchDeviceDuplicateCommandHandler : ICommandOptionsHandler<FetchDe
                 device.OsVersion
             );
         AnsiConsole.Write(table);
-        if (removeProvided && forceProviced)
+        if (removeProvided && forceProvided)
         {
-            await _deviceDuplicateService.RemoveDuplicateDevicesAsync(accessToken);
+            await _deviceService.RemoveDuplicateDevicesAsync(accessToken);
         }
 
         if (removeProvided)
