+++ IntuneAssistant.Infrastructure/Interfaces/Devices/IDeviceService.cs
@@ -3,7 +3,7 @@ using IntuneAssistant.Models.Devices;
 using IntuneAssistant.Models.Options;
 using Microsoft.Graph.Beta.Models;
 
-namespace IntuneAssistant.Infrastructure.Interfaces;
+namespace IntuneAssistant.Infrastructure.Interfaces.Devices;
 
 public interface IDeviceService
 {
@@ -11,4 +11,7 @@ public interface IDeviceService
     Task<List<ManagedDevice>?> GetNonCompliantManagedDevicesListAsync(string? accessToken);
     Task<List<ManagedDevice>?> GetFilteredDevicesListAsync(string? accessToken, DeviceFilterOptions? filterOptions);
     Task<List<OsBuildModel>> GetDevicesOsVersionsOverviewAsync(string? accessToken, DeviceFilterOptions? filterOptions);
+    
+    Task<List<ManagedDevice>?> GetDuplicateDevicesListAsync(string? accessToken, DeviceFilterOptions? filterOptions, ExportOptions? exportOptions);
+    Task<List<ManagedDevice>?> RemoveDuplicateDevicesAsync(string? accessToken);
 }
