+++ IntuneAssistant/Constants/AppConfiguration.cs
@@ -21,6 +21,7 @@ public static class AppConfiguration
     public static readonly string[] GRAPH_INTERACTIVE_SCOPE = {
         "Group.Read.All",
         "Directory.Read.All",
+        "DeviceManagementManagedDevices.Read.All",
         "DeviceManagementConfiguration.ReadWrite.All",
         "DeviceManagementServiceConfig.ReadWrite.All",
         "DeviceManagementApps.ReadWrite.All",
