+++ IntuneAssistant/Models/Devices/DeviceModel.cs
@@ -9,7 +9,8 @@ public class DeviceModel
     public string Status { get; init; } = string.Empty;
     public DateTimeOffset LastSyncDateTime { get; init; } = DateTime.Now;
     public string OsVersion { get; init; } = String.Empty;
-
+    public string ComplianceState { get; init; } = string.Empty;
+    public string UserDisplayName { get; init; } = string.Empty;
 }
 
 public static class DeviceModelExtensions
@@ -23,7 +24,9 @@ public static class DeviceModelExtensions
             DeviceName = device.DeviceName,
             Status = device.ComplianceState.ToString(),
             LastSyncDateTime = device.LastSyncDateTime.GetValueOrDefault(),
-            OsVersion = device.OsVersion
+            OsVersion = device.OsVersion,
+            UserDisplayName = device.UserDisplayName,
+            ComplianceState = device.ComplianceState.ToString()
         };
     }
 }
